ST2-777: Create Food Story (#211)

+++ CalorieCounter/Services/MealUsageService/MealUsageServiceImpl.swift
@@ -14,6 +14,11 @@ class MealUsageServiceImpl: MealUsageService {
     @Dependency(\.mealItemService) private var mealItemService
 
     func incrementUsage(_ mealItem: MealItem, mealType: MealType) async throws -> MealUsage {
+        for ingredient in mealItem.ingredients {
+            let item = try await mealItemService
+                .mealItem(byName: ingredient.mealName, creationType: ingredient.type) ?? ingredient
+            _ = try await incrementUsage(item, mealType: mealType)
+        }
         if let usage = try await mealUsageRepository.usage(byMealId: mealItem.id, type: mealType) {
             return try await mealUsageRepository.save(usage.incremented)
         }
@@ -25,6 +30,13 @@ class MealUsageServiceImpl: MealUsageService {
     }
 
     func decrementUsage(_ mealItem: MealItem, mealType: MealType) async throws {
+
+        for ingredient in mealItem.ingredients {
+            let item = try await mealItemService
+                .mealItem(byName: ingredient.mealName, creationType: ingredient.type) ?? ingredient
+            _ = try await decrementUsage(item, mealType: mealType)
+        }
+
         let usages = try await mealUsageRepository.usage(byMealId: mealItem.id)
 
         guard let targetUsage = usages.first(where: { $0.mealId == mealItem.id && $0.mealType == mealType }) else {
@@ -38,14 +50,19 @@ class MealUsageServiceImpl: MealUsageService {
 
         try await mealUsageRepository.delete(byId: decrementedUsage.id)
 
+        guard mealItem.canBeDeletedOnZeroUsage else {
+            return
+        }
         let otherUsages = usages.filter { $0.id != decrementedUsage.id }
         if otherUsages.isEmpty {
             try await mealItemService.delete(byId: mealItem.id)
         }
     }
 
-    func favoriteMealItems(count: Int, mealType: MealType) async throws -> [MealItem] {
-        let usages = try await sortedUsage(count: count, mealType: mealType)
+    func favoriteMealItems(from mealItemIds: [String],
+                           count: Int,
+                           mealType: MealType) async throws -> [MealItem] {
+        let usages = try await mealUsageRepository.sortedUsage(count: count, mealType: mealType, from: mealItemIds)
         var result: [MealItem] = []
         for usage in usages {
             if let mealItem = try await mealItemService.mealItem(byId: usage.mealId) {
@@ -55,7 +72,11 @@ class MealUsageServiceImpl: MealUsageService {
         return result
     }
 
-    func sortedUsage(count: Int, mealType: MealType) async throws -> [MealUsage] {
-        try await mealUsageRepository.sortedUsage(count: count, mealType: mealType)
+    func delete(byMealItemId id: String) async throws {
+        let usages = try await mealUsageRepository.usage(byMealId: id)
+        for usage in usages {
+            try await mealUsageRepository.delete(byId: usage.id)
+            await Task.yield()
+        }
     }
 }
